Isolate split screen composition from engine facade
split_screen is now separate from the remaining engine facade while preserving mcp_video.engine.split_screen for public callers. The module keeps historical layout behavior and sanitizes computed dimensions before FFmpeg filter interpolation.

Constraint: Keep split-screen separate from overlay and composite timeline paths to avoid overlapping filter graph changes.

Rejected: Validate unknown layouts in this PR | the engine historically treats non-side-by-side values as top-bottom, and changing that behavior belongs in a separate API compatibility change.

Confidence: high

Scope-risk: moderate

Directive: Do not combine split-screen with timeline/composite overlay refactors; keep graph-building changes reviewable.

Tested: ruff check mcp_video/engine.py mcp_video/engine_split_screen.py; ruff format --check mcp_video/engine.py mcp_video/engine_split_screen.py; python3 -m pytest tests/test_engine_advanced.py::TestSplitScreen -q --tb=short; python3 -m pytest tests/test_cli.py -k 'split_screen' tests/test_server.py -k 'split_screen' tests/test_red_team.py -k 'split_screen' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
